@@ -1,4 +1,4 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable, PLATFORM_ID } from '@angular/core';
 import { Observable } from 'rxjs/Observable';
 import { Subject } from 'rxjs/Subject';
 import { AngularFireLiteApp } from '../core.service';
@@ -6,6 +6,7 @@ import { fromPromise } from 'rxjs/observable/fromPromise';
 
 import { storage } from 'firebase/app';
 import 'firebase/storage';
+import { isPlatformBrowser } from '@angular/common';
 
 
 @Injectable()
@@ -13,7 +14,8 @@ export class AngularFireLiteStorage {
 
   private readonly storage: storage.Storage;
 
-  constructor(private app: AngularFireLiteApp) {
+  constructor(private app: AngularFireLiteApp,
+              @Inject(PLATFORM_ID) private platformId: Object) {
     this.storage = app.instance().storage();
   }
 
@@ -22,49 +24,63 @@ export class AngularFireLiteStorage {
   // ------------- Upload -----------------//
 
   upload(ref: string, file: File | Blob | Uint8Array, metadata?: Object): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).put(file, metadata));
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).put(file, metadata));
+    }
   }
 
   uploadString(ref: string, string: string): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).putString(string));
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).putString(string));
+    }
   }
 
   // ------------- Download -----------------//
 
   download(ref: string): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).getDownloadURL());
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).getDownloadURL());
+    }
   }
 
 
   // ------------- Delete -----------------//
 
   remove(ref): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).delete());
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).delete());
+    }
   }
 
   // ------------- Metadata -----------------//
 
   getMetadata(ref: string): Subject<any> {
-    const META = new Subject();
-    this.storage.ref().child(ref).getMetadata().then((meta) => {
-      META.next(meta);
-    });
-    return META;
+    if (isPlatformBrowser(this.platformId)) {
+      const META = new Subject();
+      this.storage.ref().child(ref).getMetadata().then((meta) => {
+        META.next(meta);
+      });
+      return META;
+    }
   }
 
 
   updateMetadata(ref: string, metadata: Object): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).updateMetadata(metadata));
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).updateMetadata(metadata));
+    }
   }
 
   deleteMetadata(ref: string): Observable<any> {
-    return fromPromise(this.storage.ref().child(ref).updateMetadata({
-      customMetadata: null,
-      cacheControl: null,
-      contentEncoding: null,
-      contentLanguage: null,
-      contentType: null
-    }));
+    if (isPlatformBrowser(this.platformId)) {
+      return fromPromise(this.storage.ref().child(ref).updateMetadata({
+        customMetadata: null,
+        cacheControl: null,
+        contentEncoding: null,
+        contentLanguage: null,
+        contentType: null
+      }));
+    }
   }
 
 }
